major-ish updates
split some large functions into smaller ones. this is the case with the convert function and the remove function.

also leaving only one blank line around class methods, per PEP8.

just found out you can get a return value from a function without first assigning it to a variable. that's the change here with get_timestamp. we can now insert the function right into the string.

moving some if statements inside their actual functions, instead of running the statements before calling the functions. cleaner this way.

+++ show.py
@@ -6,7 +6,6 @@ It's best to read the docs.
 © Ben Weddle is to blame for this code. Anyone is free to use it.
 '''
 
-from warnings import catch_warnings
 import xml.etree.ElementTree as ET
 import subprocess
 from datetime import datetime
@@ -60,32 +59,29 @@ class TLShow:
     
     def __str__(self) -> str:
         return "This is a really cool, useful thing. Calling this should give useful info. I'll come back to it. TODO"
-
-
-    def convert(self, input):
-        '''convert file with ffmpeg and proceed'''
+    
+    def create_output_filename(self):
         if self.include_date:
             date = datetime.now().strftime("%m%d%y")
             outputFile = (f"{self.show_filename}-{date}.wav")
         else:
             outputFile = (f'{self.show_filename}.wav')
-        TLShow.syslog(self, message=f'{self.show}: Converting to TL format.')
+        return outputFile
+
+    def convert(self, input):
+        '''convert file with ffmpeg and proceed'''
+        outputFile = TLShow.create_output_filename(self)
+        TLShow.syslog(self, message=f'{self.show}: Attempting to convert to TL format.')
 
         if self.breakaway:
             subprocess.run(f'ffmpeg -hide_banner -loglevel quiet -i {input} -ar 44100 -ac 1 -t {self.breakaway} -af loudnorm=I=-{self.ff_level} -y {outputFile}')
         else:
             subprocess.run(f'ffmpeg -hide_banner -loglevel quiet -i {input} -ar 44100 -ac 1 -af loudnorm=I=-{self.ff_level} -y {outputFile}')
-        if self.check_if_below or self.check_if_above:
-            TLShow.check_length(self, fileToCheck=outputFile) # call this before removing the files
-        else:
-            TLShow.syslog(self, message=f'{self.show}: The check length function is turned off.')
-        if self.url:
-            TLShow.remove(self, fileToDelete=input)
-        elif self.local_file:
-            if self.remove_source == True:
-                TLShow.remove(self, fileToDelete=input)
-        TLShow.copy(self, fileToCopy=outputFile)
 
+        TLShow.syslog(self, message=f'{self.show}: File converted')
+        TLShow.check_length(self, fileToCheck=outputFile) #call this before removing the file!
+        TLShow.remove(self, fileToDelete=input)
+        TLShow.copy(self, fileToCopy=outputFile)
 
     def copy(self, fileToCopy):
         '''TODO explain'''
@@ -95,19 +91,30 @@ class TLShow:
             TLShow.syslog(self, message=f'{self.show}: Copying {fileToCopy} to {destinations[numberOfDestinations]}...')
             shutil.copy(fileToCopy, destinations[numberOfDestinations])
             numberOfDestinations = numberOfDestinations-1
-        TLShow.remove(self, fileToDelete=fileToCopy)
+        #this is the file we're copying, so it is the file already converted. we always want to remove this.
+        TLShow.remove(self, fileToDelete=fileToCopy, is_output_file=True)
         TLShow.check_file_transferred(self, fileToCheck=fileToCopy)
 
+    def decide_whether_to_remove(self):
+        if self.url:
+            return True
+        elif self.local_file:
+            if self.remove_source == True:
+                return True 
 
-    def remove(self, fileToDelete):
+    def remove(self, fileToDelete, is_output_file=False):
         '''TODO explain'''
-        TLShow.syslog(self, message=f'{self.show}: Deleting {fileToDelete}')
-        os.remove(fileToDelete)  # remove original file from current directory
-
+        if TLShow.decide_whether_to_remove(self) or is_output_file:
+            TLShow.syslog(self, message=f'{self.show}: Deleting {fileToDelete}')
+            os.remove(fileToDelete)  # remove original file from current directory
 
     def removeYesterdayFiles(self):
         '''delete yesterday's files from destinations. 
         OK to use glob wildcards since there should only ever be one file'''
+        
+        def __str__(self) -> str:
+            return "This is a really cool, useful thing. Calling this should give useful info. I'll come back to it. TODO"
+
         if self.remove_yesterday:
             numberOfDestinations = len(destinations)
             numberOfDestinations = numberOfDestinations - 1
@@ -123,7 +130,6 @@ class TLShow:
                     TLShow.syslog(self, message=f"{self.show}: Cannot find yesterday's files in {destinations[numberOfDestinations]}. Continuing...")
                 numberOfDestinations = numberOfDestinations - 1
 
-
     def download_file(self, i=0):
         '''download audio file from rss feed or permalink'''
         if self.is_permalink:
@@ -135,7 +141,6 @@ class TLShow:
         # using wget because urlretrive is getting a 403 denied error
         subprocess.run(f'wget -q -O {input_file} {download}')
         TLShow.check_downloaded_file(self, fileToCheck=input_file, i=i)
-
 
     def check_downloaded_file(self, fileToCheck, i):
         '''TODO explain'''
@@ -154,16 +159,13 @@ class TLShow:
         if is_not_empty == True:
             pass
         else:
-            timestamp = get_timestamp()
             toSend = (f"There was a problem with {self.show}.\n\n\
 It looks like the downloaded file is empty. Please check manually! \
 Yesterday's file will remain.\n\n\
-{timestamp}")
+{get_timestamp()}")
             TLShow.notify(self, message=toSend, subject='Error')
             TLShow.remove(self, fileToDelete=fileToCheck)
             
-
-
     def syslog(self, message):
         '''send message to syslog server'''
         host = os.environ["syslog_server"]  # IP of PC with syslog server software
@@ -176,7 +178,6 @@ Yesterday's file will remain.\n\n\
 
         my_logger.info(message)
         my_logger.removeHandler(handler) # don't forget this after you send the message!
-
 
     def send_mail(self, message, subject):
         '''send email to TL gmail account via relay address'''
@@ -192,7 +193,6 @@ Yesterday's file will remain.\n\n\
         mail = smtplib.SMTP(host=mail_server)
         mail.send_message(format)
         mail.quit()
-
 
     def send_sms(self, message):
         '''send sms via twilio. all info is stored in PC's environement variables'''
@@ -211,7 +211,6 @@ Yesterday's file will remain.\n\n\
             )
             message.sid
 
-
     def notify(self, message, subject):
         '''TODO: explain'''
         if self.notifications:
@@ -227,7 +226,6 @@ Yesterday's file will remain.\n\n\
         else:
             pass
 
-
     def check_file_transferred(self, fileToCheck):
         '''check if file transferred to destination(s)'''
         try:
@@ -240,41 +238,39 @@ Yesterday's file will remain.\n\n\
                 TLShow.syslog(self, message=f'{self.show}: {fileToCheck} arrived at {destinations[numberOfDestinations]}')
             TLShow.countdown(self)
         except:
-            timestamp = get_timestamp()
             toSend = (f"There was a problem with {self.show}.\n\n\
 It looks like the file either wasn't converted or didn't transfer correctly. \
 Please check manually!\n\n\
-{timestamp}")
+{get_timestamp()}")
             TLShow.notify(self, message=toSend, subject='Error')
             os.system('cls')
             print(toSend)  # get user's attention!
             print()
             input('(press enter to close this window)') # force user to acknowledge by closing window
 
-
     def check_length(self, fileToCheck):
         '''check length of converted file with ffprobe. if too long or short, send notification'''
-        duration = subprocess.getoutput(f"ffprobe -v error -show_entries format=duration \
-        -of default=noprint_wrappers=1:nokey=1 {fileToCheck}")
-        duration = float(duration)
-        duration = duration/60
-        duration = round(duration, 2)
+        if self.check_if_below or self.check_if_above:
+            duration = subprocess.getoutput(f"ffprobe -v error -show_entries format=duration \
+            -of default=noprint_wrappers=1:nokey=1 {fileToCheck}")
+            duration = float(duration)
+            duration = duration/60
+            duration = round(duration, 2)
 
-        if duration > self.check_if_above:
-            timestamp = get_timestamp()
-            toSend = (f"Today's {self.show} is {duration} minutes long! \
+            if duration > self.check_if_above:
+                toSend = (f"Today's {self.show} is {duration} minutes long! \
 Please check manually and make edits to bring it below {self.check_if_above} minutes.\n\n\
-{timestamp}")
-            TLShow.notify(self, message=toSend, subject='Check Length')
-        elif duration < self.check_if_below:
-            timestamp = get_timestamp()
-            toSend = (f"Today's {self.show} is only {duration} minutes long! \
+{get_timestamp()}")
+                TLShow.notify(self, message=toSend, subject='Check Length')
+            elif duration < self.check_if_below:
+                toSend = (f"Today's {self.show} is only {duration} minutes long! \
 This is unusual and could indicate a problem with the file. Please check manually!\n\n\
-{timestamp}")
-            TLShow.notify(self, message=toSend, subject='Check Length')
+{get_timestamp()}")
+                TLShow.notify(self, message=toSend, subject='Check Length')
+            else:
+                TLShow.syslog(self, message=f'{self.show}: File is {duration} minute(s). Continuing...')
         else:
-            TLShow.syslog(self, message=f'{self.show}: File is {duration} minute(s). Continuing...')
-
+            TLShow.syslog(self, message=f'{self.show}: The check length function is turned off.')
 
     def get_feed(self):
         '''get the feed and create an ET object'''
@@ -283,7 +279,6 @@ This is unusual and could indicate a problem with the file. Please check manuall
         rssfeed = rssfeed.text
         rssfeed = ET.fromstring(rssfeed)
         return rssfeed
-
 
     def check_feed_updated(self):
         '''TODO explain'''
@@ -296,7 +291,6 @@ This is unusual and could indicate a problem with the file. Please check manuall
                 TLShow.syslog(self, message=f'{self.show}: The feed is updated. Continuing...')
                 return True
 
-
     def get_audio_url(self):
         '''TODO: explain'''
         root = TLShow.get_feed(self)
@@ -306,7 +300,6 @@ This is unusual and could indicate a problem with the file. Please check manuall
             audio_url = audio_url.get('url')
             TLShow.syslog(self, message=f'{self.show}: Audio URL is: {audio_url}')
             return audio_url
-
 
     def check_feed_loop(self):
         '''occasionally, the first time we check the feed, it is not showing as updated.
@@ -320,7 +313,6 @@ This is unusual and could indicate a problem with the file. Please check manuall
             else:
                 time.sleep(1)
                 i = i+1
-
 
     def countdown(self):
         '''the reason for this is to give a visual cue to the user
@@ -338,7 +330,6 @@ This is unusual and could indicate a problem with the file. Please check manuall
             print(f'This window will close in {number} seconds...', end='\r')
             number = number-1
             time.sleep(1)
-
 
     def run(self):
         '''Begins to process the file'''
@@ -370,17 +361,14 @@ This is unusual and could indicate a problem with the file. Please check manuall
             if self.is_permalink:
                 TLShow.removeYesterdayFiles(self)
                 TLShow.download_file(self)
-
             elif TLShow.check_feed_loop(self) == True:
                 TLShow.removeYesterdayFiles(self)
                 TLShow.download_file(self)
-                
             else:
-                timestamp = get_timestamp()
                 toSend = (f"There was a problem with {self.show}. \n\n\
 It looks like today's file hasn't yet been posted. \
 Please check and download manually! Yesterday's file will remain.\n\n\
-{timestamp}")
+{get_timestamp()}")
                 TLShow.notify(self, message=toSend, subject='Error')
                 os.system('cls')
                 print(toSend)
@@ -391,16 +379,16 @@ Please check and download manually! Yesterday's file will remain.\n\n\
             if self.local_file:
                 TLShow.check_downloaded_file(self, fileToCheck=self.local_file, i=0)
             else:
-                timestamp = get_timestamp()
                 to_send = (f"There was a problem with {self.show}. \n\n\
 It looks like the source file doesn't exist. \
 Please check manually! Yesterday's file will remain.\n\n\
-{timestamp}")
+{get_timestamp()}")
                 TLShow.notify(self, message=to_send, subject='Error')
                 os.system('cls')
                 print(to_send)
                 print()
-                input('(press enter to close this window)') #force user to acknowledge      
+                input('(press enter to close this window)') #force user to acknowledge   
+                   
         else:
             raise Exception ('Sorry, you need to specify either a URL or local audio file. \
 Did you set is_local to True?')
